change config files path
