+++ cli/configs/config-templates.ts
